pyproject.toml
[tool.poetry]
name = "cafa5"
version = "0.1.0"
description = "Deep Learning models for CAFA5 Kaggle Competition (Protein Function Accession)"
authors = ["henriupton99 <henriupton99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.2.2"
torch = "^2.3.1"
pyyaml = "^6.0.1"
pandas = "^2.2.2"
tqdm = "^4.66.4"
fastparquet = "^2024.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
